Add Iconify JSON packages and update blog profile details
This commit adds the `@iconify-json/logos` and `@iconify-json/simple-icons` packages to the project dependencies, enhancing the icon options available for use. Additionally, it updates the blog profile card to change the profile image and name, as well as adds new skills to the profile description, improving the overall presentation and information provided to users.